requirements.txt
Flask==3.0.3
python-dateutil==2.9.0
tomli==2.0.1
Pillow==10.4.0
PyPDF2==3.0.1
python-docx==1.1.2
openpyxl==3.1.5
mysql-connector-python==9.0.0
python-pptx==0.6.23
rarfile==4.2
py7zr==0.21.0
